rename vss envs and end episode when goal
