--- envs/rc_gym/vss/env_3v3/__init__.py
@@ -1,2 +1,0 @@
-from rc_gym.vss.env_3v3.vss_gym_3v3 import VSS3v3Env
-from rc_gym.vss.env_3v3.vss_gym_3v3_fira import VSS3v3FIRAEnv
