+++ envs/rc_gym/vss/env_vss/__init__.py
@@ -1,0 +1,2 @@
+from rc_gym.vss.env_vss.vss_gym import VSSEnv
+from rc_gym.vss.env_vss.vss_gym_fira import VSSFIRAEnv
