+++ envs/rc_gym/vss/env_vss/vss_gym_fira.py
@@ -6,10 +6,10 @@ from typing import Dict
 import gym
 import numpy as np
 from rc_gym.Entities import Frame, Robot
-from rc_gym.vss.vss_gym_base import VSSBaseEnv
+from rc_gym.vss.vss_gym_base import VSSBaseFIRAEnv
 
 
-class VSS3v3Env(VSSBaseEnv):
+class VSSFIRAEnv(VSSBaseFIRAEnv):
     """This environment controls a single robot in a VSS soccer League 3v3 match 
 
 
@@ -55,10 +55,19 @@ class VSS3v3Env(VSSBaseEnv):
         super().__init__(field_type=0, n_robots_blue=3, n_robots_yellow=3,
                          time_step=0.025)
 
+        low_obs_bound = [-1.2, -1.2, -1.25, -1.25]
+        low_obs_bound += [-1.2, -1.2, -1, -1, -1.25, -1.25, -1.2]*3
+        low_obs_bound += [-1.2, -1.2, -1.25, -1.25, -1.2]*3
+        high_obs_bound = [1.2, 1.2, 1.25, 1.25]
+        high_obs_bound += [1.2, 1.2, 1, 1, 1.25, 1.25, 1.2]*3
+        high_obs_bound += [1.2, 1.2, 1.25, 1.25, 1.2]*3
+        low_obs_bound = np.array(low_obs_bound, dtype=np.float32)
+        high_obs_bound = np.array(high_obs_bound, dtype=np.float32)
+
         self.action_space = gym.spaces.Box(low=-1, high=1,
                                            shape=(2, ), dtype=np.float32)
-        self.observation_space = gym.spaces.Box(low=-self.NORM_BOUNDS,
-                                                high=self.NORM_BOUNDS,
+        self.observation_space = gym.spaces.Box(low=low_obs_bound,
+                                                high=high_obs_bound,
                                                 shape=(40, ), dtype=np.float32)
 
         # Initialize Class Atributes
@@ -78,7 +87,6 @@ class VSS3v3Env(VSSBaseEnv):
     def reset(self):
         self.actions = None
         self.reward_shaping_total = None
-        self.previous_ball_potential = None
         for ou in self.ou_actions:
             ou.reset()
 
@@ -119,29 +127,29 @@ class VSS3v3Env(VSSBaseEnv):
                 self.norm_w(self.frame.robots_yellow[i].v_theta)
             )
 
-        return np.array(observation, dtype=np.float32)
+        return np.array(observation)
 
     def _get_commands(self, actions):
         commands = []
         self.actions = {}
 
         self.actions[0] = actions
-        v_wheel0, v_wheel1 = self._actions_to_v_wheels(actions)
-        commands.append(Robot(yellow=False, id=0, v_wheel0=v_wheel0,
-                              v_wheel1=v_wheel1))
+        v_wheel1, v_wheel2 = self._actions_to_v_wheels(actions)
+        commands.append(Robot(yellow=False, id=0, v_wheel1=v_wheel1,
+                              v_wheel2=v_wheel2))
 
         # Send random commands to the other robots
         for i in range(1, self.n_robots_blue):
             actions = self.ou_actions[i].sample()
             self.actions[i] = actions
-            v_wheel0, v_wheel1 = self._actions_to_v_wheels(actions)
-            commands.append(Robot(yellow=False, id=i, v_wheel0=v_wheel0,
-                                  v_wheel1=v_wheel1))
+            v_wheel1, v_wheel2 = self._actions_to_v_wheels(actions)
+            commands.append(Robot(yellow=False, id=i, v_wheel1=v_wheel1,
+                                  v_wheel2=v_wheel2))
         for i in range(self.n_robots_yellow):
             actions = self.ou_actions[self.n_robots_blue+i].sample()
-            v_wheel0, v_wheel1 = self._actions_to_v_wheels(actions)
-            commands.append(Robot(yellow=True, id=i, v_wheel0=v_wheel0,
-                                  v_wheel1=v_wheel1))
+            v_wheel1, v_wheel2 = self._actions_to_v_wheels(actions)
+            commands.append(Robot(yellow=True, id=i, v_wheel1=v_wheel1,
+                                  v_wheel2=v_wheel2))
 
         return commands
 
@@ -150,9 +158,9 @@ class VSS3v3Env(VSSBaseEnv):
         Difference of potential of the ball in time_step seconds.
         '''
         # Calculate ball potential
-        length_cm = self.field.length * 100
-        half_lenght = (self.field.length / 2.0)\
-            + self.field.goal_depth
+        length_cm = self.field_params['field_length'] * 100
+        half_lenght = (self.field_params['field_length'] / 2.0)\
+            + self.field_params['goal_depth']
 
         # distance to defence
         dx_d = (half_lenght + self.frame.ball.x) * 100
@@ -170,7 +178,7 @@ class VSS3v3Env(VSSBaseEnv):
         if self.previous_ball_potential is not None:
             diff = ball_potential - self.previous_ball_potential
             grad_ball_potential = np.clip(diff * 3 / self.time_step,
-                                          -5.0, 5.0)
+                                          -1.0, 1.0)
 
         self.previous_ball_potential = ball_potential
 
@@ -193,16 +201,16 @@ class VSS3v3Env(VSSBaseEnv):
 
         move_reward = np.dot(robot_ball, robot_vel)
 
-        move_reward = np.clip(move_reward / 0.4, -5.0, 5.0)
+        move_reward = np.clip(move_reward / 0.4, -1.0, 1.0)
         return move_reward
 
     def __energy_penalty(self):
         '''Calculates the energy penalty'''
 
-        en_penalty_1 = abs(self.sent_commands[0].v_wheel0)
-        en_penalty_2 = abs(self.sent_commands[0].v_wheel1)
+        en_penalty_1 = abs(self.sent_commands[0].v_wheel1)
+        en_penalty_2 = abs(self.sent_commands[0].v_wheel2)
         energy_penalty = - (en_penalty_1 + en_penalty_2)
-        energy_penalty /= self.field.rbt_wheel_radius
+        energy_penalty /= self.rsim.robot_wheel_radius
         return energy_penalty
 
     def _calculate_reward_and_done(self):
@@ -217,12 +225,12 @@ class VSS3v3Env(VSSBaseEnv):
                                          'goals_blue': 0, 'goals_yellow': 0}
 
         # Check if goal ocurred
-        if self.frame.ball.x > (self.field.length / 2):
+        if self.frame.ball.x > (self.field_params['field_length'] / 2):
             self.reward_shaping_total['goal_score'] += 1
             self.reward_shaping_total['goals_blue'] += 1
             reward = 10
             goal = True
-        elif self.frame.ball.x < -(self.field.length / 2):
+        elif self.frame.ball.x < -(self.field_params['field_length'] / 2):
             self.reward_shaping_total['goal_score'] -= 1
             self.reward_shaping_total['goals_yellow'] += 1
             reward = -10
@@ -251,15 +259,16 @@ class VSS3v3Env(VSSBaseEnv):
             initial_pos_frame: Frame = self._get_initial_positions_frame()
             self.rsim.reset(initial_pos_frame)
             self.frame = self.rsim.get_frame()
-            self.previous_ball_potential = None
             self.last_frame = None
 
-        return reward, False
+        done = self.steps
+
+        return reward, done
 
     def _get_initial_positions_frame(self):
-        '''Returns the position of each robot and ball for the initial frame'''
-        field_half_length = self.field.length / 2
-        field_half_width = self.field.width / 2
+        '''Returns the position of each robot and ball for the inicial frame'''
+        field_half_length = self.field_params['field_length'] / 2
+        field_half_width = self.field_params['field_width'] / 2
 
         def x(): return random.uniform(-field_half_length + 0.1,
                                        field_half_length - 0.1)
@@ -283,29 +292,35 @@ class VSS3v3Env(VSSBaseEnv):
 
         for i in range(self.n_robots_yellow):
             pos_frame.robots_yellow[i] = Robot(x=x(), y=y(), theta=theta())
-            agents.append(pos_frame.robots_yellow[i])
+            agents.append(pos_frame.robots_blue[i])
 
-        def same_position_ref(obj, ref, radius):
-            if obj.x >= ref.x - radius and obj.x <= ref.x + radius and \
-                    obj.y >= ref.y - radius and obj.y <= ref.y + radius:
+        def same_position_ref(x, y, x_ref, y_ref, radius):
+            if x >= x_ref - radius and x <= x_ref + radius and \
+                    y >= y_ref - radius and y <= y_ref + radius:
                 return True
             return False
 
-        radius_ball = self.field.ball_radius
-        radius_robot = self.field.rbt_radius
+        radius_ball = 0.2
+        radius_robot = 0.2
+        same_pos = True
 
-        for i in range(len(agents)):
-            while same_position_ref(agents[i], pos_frame.ball, radius_ball):
-                agents[i] = Robot(x=x(), y=y(), theta=theta())
-            for j in range(i):
-                while same_position_ref(agents[i], agents[j], radius_robot):
+        while same_pos:
+            for i in range(len(agents)):
+                same_pos = False
+                while same_position_ref(agents[i].x, agents[i].y, pos_frame.ball.x, pos_frame.ball.y, radius_ball):
                     agents[i] = Robot(x=x(), y=y(), theta=theta())
+                    same_pos = True
+                for j in range(i + 1, len(agents)):
+                    while same_position_ref(agents[i].x, agents[i].y, agents[j].x, agents[j].y, radius_robot):
+                        agents[i] = Robot(x=x(), y=y(), theta=theta())
+                        same_pos = True
 
-        for i in range(self.n_robots_blue):
-            pos_frame.robots_blue[i] = agents[i]
-
-        for i in range(self.n_robots_yellow):
-            pos_frame.robots_yellow[i] = agents[i+self.n_robots_blue]
+        pos_frame.robots_blue[0] = agents[0]
+        pos_frame.robots_blue[1] = agents[1]
+        pos_frame.robots_blue[2] = agents[2]
+        pos_frame.robots_yellow[0] = agents[3]
+        pos_frame.robots_yellow[1] = agents[4]
+        pos_frame.robots_yellow[2] = agents[5]
 
         return pos_frame
